Add Aloha::SlackUser to fix bug with new users
The problem:

When an existing user logs in, the data object contains
a `user` attribute. Like so:

```
data.user
```

And that can be used as a key to pull a user out of `client.users`.

But, when a user first joins a team, data is different

```
data.user
```

And that _can't_ be used in the same way. If you try

```
client.users[data.user]
```

you'll get nil. then you'll get a lot of NoMethodErrors

The Fix

Encapsulate all user finding in a new `SlackUser` object. Named this way
to differentiate it from the already extant `User` and to indicate where
the user data is coming from.

All other code that found users now uses `SlackUser`