Create users as they log in
User model records are only created when the New User hook is fired. And
messages are only delivered to users that exist in the database.

If, at some point, we want to add some onboarding message that goes to
all users, not just new users, then those users should exist in the db.